Make HttpIntegration the base class, HttpApi = buffered, HttpStreaming = streamed
See https://github.com/xp-forge/lambda-ws/pull/10#issuecomment-1819689077
